Check version number, error handling and refactoring
