+++ src/D2L.Bmx/UpdateChecker.cs
@@ -1,7 +1,6 @@
 using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Text.Json;
-using System.Text.Json.Serialization;
 
 namespace D2L.Bmx;
 
@@ -12,7 +11,7 @@ internal static class UpdateChecker {
 			var localVersion = Assembly.GetExecutingAssembly().GetName().Version;
 			var latestVersion = new Version( cachedVersion?.VersionName ?? "0.0.0" );
 			if( ShouldFetchLatestVersion( cachedVersion ) ) {
-				latestVersion = new Version( await GetLatestReleaseVersionAsync() );
+				latestVersion = new Version( GetLatestReleaseVersion( await GetLatestReleaseDataAsync() ) );
 			}
 
 			string updateLocation = string.Equals( config.Org, "d2l", StringComparison.OrdinalIgnoreCase )
@@ -49,7 +48,13 @@ internal static class UpdateChecker {
 		Console.Error.WriteLine();
 	}
 
-	private static async Task<string> GetLatestReleaseVersionAsync() {
+	internal static string GetLatestReleaseVersion( GithubRelease? releaseData ) {
+		string version = releaseData?.TagName?.TrimStart( 'v' ) ?? string.Empty;
+		SaveLatestVersion( version );
+		return version;
+	}
+
+	internal static async Task<GithubRelease?> GetLatestReleaseDataAsync() {
 		using var httpClient = new HttpClient();
 		httpClient.BaseAddress = new Uri( "https://api.github.com" );
 		httpClient.Timeout = TimeSpan.FromSeconds( 2 );
@@ -62,9 +67,7 @@ internal static class UpdateChecker {
 			responseStream,
 			SourceGenerationContext.Default.GithubRelease
 		);
-		string version = releaseData?.TagName?.TrimStart( 'v' ) ?? string.Empty;
-		SaveLatestVersion( version );
-		return version;
+		return releaseData;
 	}
 
 	private static void SaveLatestVersion( string version ) {
@@ -112,24 +115,6 @@ internal static class UpdateChecker {
 	}
 }
 
-internal record GithubRelease {
-	[JsonPropertyName( "tag_name" )]
-	public string? TagName { get; set; }
-
-	[JsonPropertyName( "assets" )]
-	public List<Assets>? Assets { get; set; }
-}
-
-internal record Assets {
-	[JsonPropertyName( "url" )]
-	public string? Url { get; set; }
-
-	[JsonPropertyName( "name" )]
-	public string? Name { get; set; }
-
-	[JsonPropertyName( "browser_download_url" )]
-	public string? BrowserDownloadUrl { get; set; }
-}
 
 internal record UpdateCheckCache {
 	public string? VersionName { get; set; }
